Rename AdditionalActionsDropdown -> ThreeDotsMenu
